feat: Acl 설정을 위한 스켈레톤 추가

+++ src/main/kotlin/com/soonyong/customdeck/server/domain/button/model/Button.kt
@@ -1,6 +1,6 @@
 package com.soonyong.customdeck.server.domain.button.model
 
 
-data class Button(val id: Long, var name: String = "button $id") {
+data class Button(val id: Int, var name: String = "button $id") {
 
 }
